RAB-1336: fix command declaration, move all the process in application
