+++ src/Akeneo/Platform/Installer/back/src/Application/DatabaseInstall/DatabaseInstallHandler.php
@@ -7,18 +7,20 @@ declare(strict_types=1);
  * @license   https://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
 
-namespace Akeneo\Platform\Installer\Application\ResetElasticsearchIndexes;
+namespace Akeneo\Platform\Installer\Application\DatabaseInstall;
 
 use Akeneo\Platform\Installer\Domain\Query\Elasticsearch\ResetIndexesInterface;
 
-final class ResetElasticSearchIndexesHandler
+final class DatabaseInstallHandler
 {
     public function __construct(
         private readonly ResetIndexesInterface $resetIndexes
     ) {}
-    public function handle(ResetElasticSearchIndexesCommand $elasticSearchIndexesCommand): void
+
+    public function handle(DatabaseInstallCommand $command): void
     {
-        $elasticSearchIndexesCommand->getIo()->info('Reset elasticsearch indexes');
-        $this->resetIndexes->reset();
+        $io = $command->getIo();
+
+        $io->title('Prepare database schema');
     }
 }
